+++ src/snowflake/snowpark/_internal/analyzer/snowflake_plan.py
@@ -53,6 +53,7 @@ from snowflake.snowpark._internal.analyzer.analyzer_utils import (
     merge_statement,
     pivot_statement,
     project_statement,
+    rename_statement,
     result_scan_statement,
     sample_statement,
     schema_cast_named,
@@ -662,6 +663,18 @@ class SnowflakePlanBuilder:
     ) -> SnowflakePlan:
         return self.build(
             lambda x: unpivot_statement(value_column, name_column, column_list, x),
+            child,
+            source_plan,
+        )
+
+    def rename(
+        self,
+        column_map: Dict[str, str],
+        child: SnowflakePlan,
+        source_plan: Optional[LogicalPlan],
+    ) -> SnowflakePlan:
+        return self.build(
+            lambda x: rename_statement(column_map, x),
             child,
             source_plan,
         )