+++ src/snowflake/snowpark/_internal/type_utils.py
@@ -43,6 +43,8 @@ from snowflake.snowpark.types import (
     FloatType,
     Geography,
     GeographyType,
+    Geometry,
+    GeometryType,
     IntegerType,
     LongType,
     MapType,
@@ -87,6 +89,8 @@ def convert_sf_to_sp_type(
         return MapType(StringType(), StringType())
     if column_type_name == "GEOGRAPHY":
         return GeographyType()
+    if column_type_name == "GEOMETRY":
+        return GeometryType()
     if column_type_name == "BOOLEAN":
         return BooleanType()
     if column_type_name == "BINARY":
@@ -173,6 +177,8 @@ def convert_sp_to_sf_type(datatype: DataType) -> str:
         return "VARIANT"
     if isinstance(datatype, GeographyType):
         return "GEOGRAPHY"
+    if isinstance(datatype, GeometryType):
+        return "GEOMETRY"
     raise TypeError(f"Unsupported data type: {datatype.__class__.__name__}")
 
 
@@ -480,6 +486,9 @@ def python_type_to_snow_type(tp: Union[str, Type]) -> Tuple[DataType, bool]:
     if tp == Geography:
         return GeographyType(), False
 
+    if tp == Geometry:
+        return GeometryType(), False
+
     raise TypeError(f"invalid type {tp}")
 
 
@@ -495,6 +504,7 @@ def snow_type_to_dtype_str(snow_type: DataType) -> str:
             TimestampType,
             TimeType,
             GeographyType,
+            GeometryType,
             VariantType,
         ),
     ):