+++ src/snowflake/snowpark/dataframe.py
@@ -81,6 +81,7 @@ from snowflake.snowpark._internal.analyzer.unary_plan_node import (
     LocalTempView,
     PersistedView,
     Project,
+    Rename,
     Sample,
     Sort,
     Unpivot,
@@ -1576,7 +1577,7 @@ class DataFrame:
             ... (1, 3000, 'FEB'),
             ... (2, 200, 'FEB') ''').collect()
             >>> df = session.table("monthly_sales")
-            >>> df.pivot("month", ['JAN', 'FEB']).sum("amount").show()
+            >>> df.pivot("month", ['JAN', 'FEB']).sum("amount").sort(df["empid"]).show()
             -------------------------------
             |"EMPID"  |"'JAN'"  |"'FEB'"  |
             -------------------------------
@@ -3365,6 +3366,78 @@ class DataFrame:
         return res_df
 
     @df_api_usage
+    def rename(
+        self,
+        col_or_mapper: Union[ColumnOrName, dict],
+        new_column: str = None,
+    ):
+        """
+        Returns a DataFrame with the specified column ``col_or_mapper`` renamed as ``new_column``. If ``col_or_mapper``
+        is a dictionary, multiple columns will be renamed in the returned DataFrame.
+
+        Example::
+            >>> # This example renames the column `A` as `NEW_A` in the DataFrame.
+            >>> df = session.sql("select 1 as A, 2 as B")
+            >>> df_renamed = df.rename(col("A"), "NEW_A")
+            >>> df_renamed.show()
+            -----------------
+            |"NEW_A"  |"B"  |
+            -----------------
+            |1        |2    |
+            -----------------
+            <BLANKLINE>
+            >>> # This example renames the column `A` as `NEW_A` and `B` as `NEW_B` in the DataFrame.
+            >>> df = session.sql("select 1 as A, 2 as B")
+            >>> df_renamed = df.rename({col("A"): "NEW_A", "B":"NEW_B"})
+            >>> df_renamed.show()
+            ---------------------
+            |"NEW_A"  |"NEW_B"  |
+            ---------------------
+            |1        |2        |
+            ---------------------
+            <BLANKLINE>
+
+        Args:
+            col_or_mapper: The old column instance or column name to be renamed, or the dictionary mapping from column instances or columns names to their new names (string)
+            new_column: The new column name (string value), if a single old column is given
+        """
+        if new_column is not None:
+            return self.with_column_renamed(col_or_mapper, new_column)
+
+        if not isinstance(col_or_mapper, dict):
+            raise ValueError(
+                f"If new_column parameter is not specified, col_or_mapper needs to be of type dict, "
+                f"not {type(col_or_mapper).__name__}"
+            )
+
+        if len(col_or_mapper) == 0:
+            raise ValueError("col_or_mapper dictionary cannot be empty")
+
+        column_or_name_list, rename_list = zip(*col_or_mapper.items())
+        for name in rename_list:
+            if not isinstance(name, str):
+                raise TypeError(
+                    f"You cannot rename a column using value {name} of type {type(name).__name__} as it "
+                    f"is not a string."
+                )
+
+        names = self._get_column_names_from_column_or_name_list(column_or_name_list)
+        normalized_name_list = [quote_name(n) for n in names]
+        rename_map = {k: v for k, v in zip(normalized_name_list, rename_list)}
+        rename_plan = Rename(rename_map, self._plan)
+
+        if self._select_statement:
+            return self._with_plan(
+                SelectStatement(
+                    from_=SelectSnowflakePlan(
+                        rename_plan, analyzer=self._session._analyzer
+                    ),
+                    analyzer=self._session._analyzer,
+                )
+            )
+        return self._with_plan(rename_plan)
+
+    @df_api_usage
     def with_column_renamed(self, existing: ColumnOrName, new: str) -> "DataFrame":
         """Returns a DataFrame with the specified column ``existing`` renamed as ``new``.
 
@@ -3384,8 +3457,6 @@ class DataFrame:
         Args:
             existing: The old column instance or column name to be renamed.
             new: The new column name.
-
-        :meth:`with_column_renamed` is an alias of :meth:`rename`.
         """
         new_quoted_name = quote_name(new)
         if isinstance(existing, str):
@@ -3408,7 +3479,7 @@ class DataFrame:
                     f"Unable to rename column {existing} because it doesn't exist."
                 )
         else:
-            raise TypeError("'exisitng' must be a column name or Column object.")
+            raise TypeError(f"{str(existing)} must be a column name or Column object.")
 
         to_be_renamed = [x for x in self._output if x.name.upper() == old_name.upper()]
         if not to_be_renamed:
@@ -3650,6 +3721,36 @@ Query List:
         df._statement_params = self._statement_params
         return df
 
+    def _get_column_names_from_column_or_name_list(
+        self, exprs: List[ColumnOrName]
+    ) -> List:
+        names = []
+        for c in exprs:
+            if isinstance(c, str):
+                names.append(c)
+            elif isinstance(c, Column) and isinstance(c._expression, Attribute):
+                names.append(
+                    self._plan.expr_to_alias.get(
+                        c._expression.expr_id, c._expression.name
+                    )
+                )
+            elif (
+                isinstance(c, Column)
+                and isinstance(c._expression, UnresolvedAttribute)
+                and c._expression.df_alias
+            ):
+                names.append(
+                    self._plan.df_aliased_col_name_to_real_col_name.get(
+                        c._expression.name, c._expression.name
+                    )
+                )
+            elif isinstance(c, Column) and isinstance(c._expression, NamedExpression):
+                names.append(c._expression.name)
+            else:
+                raise TypeError(f"{str(c)} must be a column name or Column object.")
+
+        return names
+
     def _convert_cols_to_exprs(
         self,
         calling_method: str,
@@ -3708,6 +3809,3 @@ Query List:
     # joinTableFunction = join_table_function
     # naturalJoin = natural_join
     # withColumns = with_columns
-
-    # Add this alias because snowpark scala has rename
-    rename = with_column_renamed