+++ tests/integ/scala/test_async_job_suite.py
@@ -287,6 +287,11 @@ def test_multiple_queries(session, resources_path):
     async_job = df.collect_nowait()
     Utils.check_answer(async_job.result(), df)
 
+    Utils.check_answer(session.create_dataframe(df.to_pandas(block=False).result()), df)
+    Utils.check_answer(
+        session.create_dataframe(next(df.to_pandas_batches(block=False).result())), df
+    )
+
     # make sure temp object is dropped
     temp_object = async_job._post_actions[0].sql.split(" ")[-1]
     with pytest.raises(SnowparkSQLException, match="does not exist or not authorized"):