+++ tests/integ/scala/test_datatype_suite.py
@@ -19,6 +19,7 @@ from snowflake.snowpark.types import (
     DoubleType,
     FloatType,
     GeographyType,
+    GeometryType,
     IntegerType,
     LongType,
     MapType,
@@ -37,7 +38,8 @@ def test_verify_datatypes_reference(session):
     schema = StructType(
         [
             StructField("var", VariantType()),
-            StructField("geo", GeographyType()),
+            StructField("geography", GeographyType()),
+            StructField("geometry", GeometryType()),
             StructField("date", DateType()),
             StructField("time", TimeType()),
             StructField("timestamp", TimestampType()),
@@ -64,6 +66,7 @@ def test_verify_datatypes_reference(session):
                 None,
                 None,
                 None,
+                None,
                 "a",
                 True,
                 None,
@@ -83,7 +86,8 @@ def test_verify_datatypes_reference(session):
 
     assert (
         str(df.schema.fields) == "[StructField('VAR', VariantType(), nullable=True), "
-        "StructField('GEO', GeographyType(), nullable=True), "
+        "StructField('GEOGRAPHY', GeographyType(), nullable=True), "
+        "StructField('GEOMETRY', GeometryType(), nullable=True), "
         "StructField('DATE', DateType(), nullable=True), "
         "StructField('TIME', TimeType(), nullable=True), "
         "StructField('TIMESTAMP', TimestampType(), nullable=True), "
@@ -125,7 +129,8 @@ def test_dtypes(session):
     schema = StructType(
         [
             StructField("var", VariantType()),
-            StructField("geo", GeographyType()),
+            StructField("geography", GeographyType()),
+            StructField("geometry", GeometryType()),
             StructField("date", DateType()),
             StructField("time", TimeType()),
             StructField("timestamp", TimestampType()),
@@ -152,6 +157,7 @@ def test_dtypes(session):
                 None,
                 None,
                 None,
+                None,
                 "a",
                 True,
                 None,
@@ -171,7 +177,8 @@ def test_dtypes(session):
 
     assert df.dtypes == [
         ("VAR", "variant"),
-        ("GEO", "geography"),
+        ("GEOGRAPHY", "geography"),
+        ("GEOMETRY", "geometry"),
         ("DATE", "date"),
         ("TIME", "time"),
         ("TIMESTAMP", "timestamp"),