+++ tests/utils.py
@@ -30,6 +30,7 @@ from snowflake.snowpark.types import (
     DecimalType,
     DoubleType,
     GeographyType,
+    GeometryType,
     LongType,
     MapType,
     StringType,
@@ -216,12 +217,13 @@ class Utils:
         actual: Union[Row, List[Row], DataFrame],
         expected: Union[Row, List[Row], DataFrame],
         sort=True,
+        statement_params=None,
     ) -> None:
         def get_rows(input_data: Union[Row, List[Row], DataFrame]):
             if isinstance(input_data, list):
                 rows = input_data
             elif isinstance(input_data, DataFrame):
-                rows = input_data.collect()
+                rows = input_data.collect(statement_params=statement_params)
             elif isinstance(input_data, Row):
                 rows = [input_data]
             else:
@@ -610,6 +612,38 @@ class TestData:
         )
 
     @classmethod
+    def geometry(cls, session: "Session") -> DataFrame:
+        return session.sql(
+            """
+            select *
+            from values
+            ('{
+                "coordinates": [
+                  30,
+                  10
+                ],
+                "type": "Point"
+            }') as T(a)
+            """
+        )
+
+    @classmethod
+    def geometry_type(cls, session: "Session") -> DataFrame:
+        return session.sql(
+            """
+            select to_geometry(a) as geo
+            from values
+            ('{
+                "coordinates": [
+                  20,
+                  81
+                ],
+                "type": "Point"
+            }') as T(a)
+            """
+        )
+
+    @classmethod
     def null_json1(cls, session: "Session") -> DataFrame:
         return session.sql(
             'select parse_json(column1) as v from values (\'{"a": null}\'), (\'{"a": "foo"}\'),'
@@ -907,4 +941,5 @@ TYPE_MAP = [
     TypeMap("object", "object", MapType(StringType(), StringType())),
     TypeMap("array", "array", ArrayType(StringType())),
     TypeMap("geography", "geography", GeographyType()),
+    TypeMap("geometry", "geometry", GeometryType()),
 ]