+++ CHANGELOG.md
@@ -2,9 +2,32 @@
 
 ## 1.6.0 (TBD)
 
+## New Features
+
+- Added support for new functions in `snowflake.snowpark.functions`:
+  - `array_sort`
+  - `sort_array`
+  - `array_min`
+  - `array_max`
+- Added support for async execution of multi-query dataframe containing binding variables.
+- Added support for renaming multiple columns in `DataFrame.rename`.
+- Added support for Geometry datatypes.
+
+### Improvements
+
+- Redundant dependency `typing-extensions` removed.
+
+### Bug Fixes
+
+- Fixed a bug where type check happens on pandas before it is imported
+
 ### Behavior Changes
 
 - `DataFrameWriter.save_as_table` now respects `nullable` field of for schema provided by the user, or inferred schema based on data from user input.
+
+### Dependency updates
+
+- Updated ``snowflake-connector-python`` to 3.0.4.
 
 ## 1.5.1 (2023-06-20)